requirements.txt
simplegmail==4.1.1
unidecode==1.3.8
pypub3>=2.0.7
ebooklib>=0.18
pillow==10.4.0
beautifulsoup4==4.12.3
